refactor : split function
